WFLY-9598 Update "xsite" tests to use CLI * remove XSiteBackupForTestCase
