+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/xsite/web.xml
